Return the semi-join of the i-table in a DT[i]-style data.table join

Write (and optionally run) data.table code to return the semi-join of the i-table in an enhanced DT[i]-style join. Arguments are as for dtjoin, except for those controlling the output columns, which do not apply.

Usage

dtjoin_semi_i(
  .DT = NULL,
  .i = NULL,
  on,
  match.na = FALSE,
  mult = "all",
  mult.DT = "all",
  nomatch = NULL,
  nomatch.DT = NULL,
  select = NULL,
  do = !(is.null(.DT) && is.null(.i)),
  show = !do,
  verbose = FALSE,
  ...
)

Arguments

.DT, .i

data.frame-like objects (plain, tibble, data.table etc.), or else both omitted (NULL) for a mock join statement with no data.

on

A character vector of join predicates, e.g. c("id", "col_DT == col_i", "date < date").

match.na

If TRUE, allow equality matches between NAs or NaNs. The default is FALSE, i.e. such matches are not allowed, as in most real-world applications (but unlike other join frameworks in R).

mult

(as in [.data.table) When a row of .i has multiple matching rows in .DT, which to accept. One of "all" (the default), "first", or "last".

mult.DT

Like mult, but with the roles of .DT and .i reversed, i.e. when a row of .DT has multiple matching rows in .i, which to accept (default "all"). Can be combined with mult.

nomatch, nomatch.DT

Permitted for consistency with dtjoin but have no effect on the resulting semi-join.

do

Whether to execute the join. Default is TRUE unless .DT and .i are both omitted/NULL, in which case a mock join statement is produced.

show

Whether to print the code for the join to the console. Default is the opposite of do. If .DT and .i are both omitted/NULL, mock join code is displayed.

verbose

(passed to [.data.table) Whether data.table should print information to the console during execution. Default FALSE.

...

Further arguments (for internal use).

Value

A data.table (the resulting semi-join), or NULL if do is FALSE. The data.table code is always printed to the console.

Examples

# TODO