fjoin

fjoin is a general-purpose data frame join package that runs on data.table and works seamlessly in general workflows and tidyverse pipelines. It is lightweight, fast, and has short and clear syntax with options not found in other frameworks.

Vignettes

Visit the package website https://trobx.github.io/fjoin or access locally in R:

Guide to fjoin: vignette("fjoin")

Performance overview: vignette("fjoin-performance")

API

fjoin_* functions	dtjoin* functions

x/y style	Extended DT[i]-like interface

fjoin_inner(), fjoin_left(), fjoin_right(), fjoin_full()	dtjoin()

fjoin_left_semi() (alias fjoin_semi()), fjoin_right_semi()	dtjoin_semi()

fjoin_left_anti() (alias fjoin_anti()), fjoin_right_anti()	dtjoin_anti()

fjoin_cross()	dtjoin_cross()

See also

Author

Maintainer: Toby Robertson toby.robertson@outlook.com